dev-requirements.txt
behaving==3.1.5
Appium-Python-Client==2.10.1
flake8==6.0.0
pytest-ckan
pytest-cov
selenium<4.10
